pyproject.toml
[tool.poetry]
name = "anime-scraper"
version = "0.1.0"
description = "Scraper for anime sites"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
m3u8 = "^3.6.0"
python-ffmpeg = "^2.0.4"
aiohttp = "^3.8.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
